requirements.txt
streamlit
psycopg2-binary
bcrypt
pandas
plotly
Pillow
werkzeug
